fixed float bug

+++ cogs/utility.py
@@ -176,7 +176,7 @@ class Utility(commands.Cog):
     @commands.command()
     async def latex(self, ctx, *, arg):
         img = self.latexToImage(arg)
-        img = img.resize((img.width * 1.5, img.height * 1.5), Image.ANTIALIAS)
+        img = img.resize((int(img.width * 1.5), int(img.height * 1.5)), Image.ANTIALIAS)
         with BytesIO() as image_binary:
             img.save(image_binary, 'PNG')
             image_binary.seek(0)
